+++ frontend/src/metabase/querying/notebook/components/JoinStep/JoinConditionColumnPicker/JoinColumnButton/index.ts
@@ -1,0 +1,1 @@
+export * from "./JoinColumnButton";
