+++ frontend/src/metabase/querying/notebook/components/JoinStep/JoinConditionColumnPicker/JoinColumnDropdown/index.ts
@@ -1,0 +1,1 @@
+export * from "./JoinColumnDropdown";
